Fix unhandled "unshift is not a function" error
In certain cases, if there is only one entry for `memberOf`, the
attribute may be a string rather than an array. Attempting to call
`unshift` on the string object would cause an unhandled error. Added a
check to ensure that `memberOf` is an array before calling `unshift`,
and if it is a string, convert it into an array with the primary group
prepended.